+++ apps/app/src/atoms/family-kind-filter.ts
@@ -1,4 +1,4 @@
-import { FamilyKind } from "entropretty-utils"
+import { FamilyKind } from "@entropretty/utils"
 import { atomWithStorage } from "jotai/utils"
 
 export const familyKindFilterAtom = atomWithStorage<FamilyKind | "All">(