+++ apps/app/src/components/AlgorithmRow/index.tsx
@@ -5,7 +5,7 @@ import { Button } from "@/components/ui/button"
 import { useAuth } from "@/contexts/auth-context"
 import { AlgorithmBitmap } from "@/features/create/AlgorithmBitmap"
 import { AlgorithmView } from "@/lib/helper.types"
-import { getSeed, seedToKey } from "entropretty-utils"
+import { getSeed, seedToKey } from "@entropretty/utils"
 import { useMemo } from "react"
 import { Link } from "react-router"