+++ apps/app/src/components/NewDialog/index.tsx
@@ -8,7 +8,7 @@ import {
   DialogTrigger,
 } from "@/components/ui/dialog"
 import { cn, familyKindColor } from "@/lib/utils"
-import { FamilyKind } from "entropretty-utils"
+import { FamilyKind } from "@entropretty/utils"
 import { ArrowUpRight, PlusIcon } from "lucide-react"
 import { Link, useNavigate } from "react-router"