+++ apps/app/src/lib/utils.ts
@@ -1,5 +1,5 @@
 import { clsx, type ClassValue } from "clsx"
-import { FamilyKind } from "entropretty-utils"
+import { FamilyKind } from "@entropretty/utils"
 import { twMerge } from "tailwind-merge"
 
 export function cn(...inputs: ClassValue[]) {