+++ packages/utils/generatePreludeString.ts
@@ -1,6 +1,21 @@
 import { writeFileSync } from "node:fs"
 import { COLORS } from "./src/prelude/colors"
-import { bit, bits, bytesToNibbles, fillEach, getByte, numeric, randomGenerator, sfc32, split, strokeEach, symmetrical } from "./src/prelude/functions"
+import {
+  bit,
+  bits,
+  bytesToNibbles,
+  fillEach,
+  getByte,
+  numeric,
+  randomGenerator,
+  sfc32,
+  split,
+  strokeEach,
+  symmetrical,
+  bytesToBits,
+  bytesToQuarterBytes,
+  map,
+} from "./src/prelude/functions"
 
 export * from "./src/prelude/colors"
 export * from "./src/prelude/functions"
@@ -22,6 +37,12 @@ export const preludeScriptString = `
   ${numeric.toString()}
   ${randomGenerator.toString()}
   ${sfc32.toString()}
+  ${bytesToBits.toString()}
+  ${bytesToQuarterBytes.toString()}
+  ${map.toString()}
 `
-writeFileSync("./src/prelude/generatedScriptString.ts", `export const preludeScriptString=\`${preludeScriptString}\``);
-console.log("Prelude script string generated");
+writeFileSync(
+  "./src/prelude/generatedScriptString.ts",
+  `export const preludeScriptString=\`${preludeScriptString}\``,
+)
+console.log("Prelude script string generated")