+++ apps/app/CHANGELOG.md
@@ -5,4 +5,4 @@
 ### Patch Changes
 
 - Updated dependencies [78c1720]
-  - entropretty-utils@1.1.0
+  - @entropretty/utils@1.1.0